Define an FFI bool as a unsigned char, so bool members in structs work

+++ src/org/jruby/ext/ffi/jffi/FFIUtil.java
@@ -27,7 +27,7 @@ public final class FFIUtil {
     private static final Map<NativeType, com.kenai.jffi.Type> buildTypeMap() {
         Map<NativeType, com.kenai.jffi.Type> m = new EnumMap<NativeType, com.kenai.jffi.Type>(NativeType.class);
         m.put(NativeType.VOID, com.kenai.jffi.Type.VOID);
-        m.put(NativeType.BOOL, com.kenai.jffi.Type.UINT32);
+        m.put(NativeType.BOOL, com.kenai.jffi.Type.UINT8);
         m.put(NativeType.CHAR, com.kenai.jffi.Type.SINT8);
         m.put(NativeType.SHORT, com.kenai.jffi.Type.SINT16);
         m.put(NativeType.INT, com.kenai.jffi.Type.SINT32);
